Moves strategy around.
