+++ src/main/java/com/scaler/lld/design/behavioural/strategy/calculator/Calculator.java
@@ -1,10 +1,10 @@
-package com.scaler.lld.design.behavioural.strategy;
+package com.scaler.lld.design.behavioural.strategy.calculator;
 
 // Step 3 - Add a field for strategy 
 public class Calculator {
 
     private OperationStrategy strategy = null;
-    
+
     public void setStrategy(OperationStrategy operationStrategy) {
         this.strategy = operationStrategy;
     }
@@ -12,5 +12,5 @@ public class Calculator {
     public int performCalculation(int a, int b) {
         return strategy.operate(a, b);
     }
-    
+
 }
